// frontend/src/api-client/models/CardInUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Simple Kanban API
 * A simple kanban API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { CardColumn } from './CardColumn';
import {
    CardColumnFromJSON,
    CardColumnFromJSONTyped,
    CardColumnToJSON,
    CardColumnToJSONTyped,
} from './CardColumn';
import type { LabelColor } from './LabelColor';
import {
    LabelColorFromJSON,
    LabelColorFromJSONTyped,
    LabelColorToJSON,
    LabelColorToJSONTyped,
} from './LabelColor';

/**
 * 
 * @export
 * @interface CardInUpdate
 */
export interface CardInUpdate {
    /**
     * 
     * @type {string}
     * @memberof CardInUpdate
     */
    title?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CardInUpdate
     */
    description?: string | null;
    /**
     * 
     * @type {CardColumn}
     * @memberof CardInUpdate
     */
    column?: CardColumn | null;
    /**
     * 
     * @type {Array<LabelColor>}
     * @memberof CardInUpdate
     */
    labels?: Array<LabelColor> | null;
}



/**
 * Check if a given object implements the CardInUpdate interface.
 */
export function instanceOfCardInUpdate(value: object): value is CardInUpdate {
    return true;
}

export function CardInUpdateFromJSON(json: any): CardInUpdate {
    return CardInUpdateFromJSONTyped(json, false);
}

export function CardInUpdateFromJSONTyped(json: any, ignoreDiscriminator: boolean): CardInUpdate {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'] == null ? undefined : json['title'],
        'description': json['description'] == null ? undefined : json['description'],
        'column': json['column'] == null ? undefined : CardColumnFromJSON(json['column']),
        'labels': json['labels'] == null ? undefined : ((json['labels'] as Array<any>).map(LabelColorFromJSON)),
    };
}

export function CardInUpdateToJSON(json: any): CardInUpdate {
    return CardInUpdateToJSONTyped(json, false);
}

export function CardInUpdateToJSONTyped(value?: CardInUpdate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'title': value['title'],
        'description': value['description'],
        'column': CardColumnToJSON(value['column']),
        'labels': value['labels'] == null ? undefined : ((value['labels'] as Array<any>).map(LabelColorToJSON)),
    };
}